refactor: split naval units panel support into scoped parts (#3878)
Decompose the 577-line naval_units_panel.dart into combine, home-transfer,
and dialog part files so each concern stays under the panel submodule size
target without behavior changes.

Refs #3878
