+++ app/lib/features/game/widgets/naval_units_panel.dart
@@ -28,6 +28,10 @@ import 'units/shared/location_section_header.dart';
 import 'units/shared/region_section_header.dart';
 import '../utils/region_labels.dart';
 
+part 'naval_units_panel_support_combine.dart';
+part 'naval_units_panel_support_home_transfer.dart';
+part 'naval_units_panel_support_dialogs.dart';
+
 class NavalUnitsPanel extends StatefulWidget with GamePanelMixin {
   const NavalUnitsPanel({
     super.key,
@@ -97,247 +101,6 @@ class _NavalUnitsPanelState extends BaseUnitsPanelState<NavalUnitsPanel> {
     super.dispose();
   }
 
-  /// Canonical fleet id for combine/split selection (Home Fleet uses [homeFleetIdFor]).
-  String _selectionFleetId(FleetRow row) {
-    if (row.isHomeFleet) return homeFleetIdFor(widget.humanPlayerId);
-    return row.fleetId;
-  }
-
-  bool _canCombineSelection(List<FleetRow> flat) {
-    final rowsById = <String, FleetRow>{
-      for (final r in flat) _selectionFleetId(r): r,
-    };
-    final activeIds = selection.selectedIds
-        .where(rowsById.containsKey)
-        .toList();
-    if (activeIds.length < 2) return false;
-    final homeTransferRows = _homeTransferRows(flat, activeIds.toSet());
-    if (homeTransferRows != null) {
-      return _isEligibleHomeTransferSource(homeTransferRows.source);
-    }
-    String? locationKey;
-    for (final id in activeIds) {
-      final row = rowsById[id]!;
-      locationKey ??= row.locationKey;
-      if (row.locationKey != locationKey) return false;
-    }
-    return true;
-  }
-
-  String _combineTargetFleetId(List<FleetRow> flat, Set<String> selected) {
-    for (final row in flat) {
-      final id = _selectionFleetId(row);
-      if (!selected.contains(id)) continue;
-      if (row.isHomeFleet) return id;
-    }
-    for (final row in flat) {
-      final id = _selectionFleetId(row);
-      if (selected.contains(id)) return id;
-    }
-    throw StateError('combine target: empty selection');
-  }
-
-  Fleet? _fleetForRow(FleetRow row) {
-    final id = _selectionFleetId(row);
-    final found = widget.game.fleetById(id);
-    if (found != null) return found;
-    if (row.isHomeFleet) {
-      final portId = row.inPortAtProvinceId;
-      if (portId == null) return null;
-      return Fleet(
-        id: id,
-        ownerId: widget.humanPlayerId,
-        regionId: row.regionId,
-        inPortAtProvinceId: portId,
-        ships: const [],
-        mission: FleetMission.none,
-      );
-    }
-    return null;
-  }
-
-  void _toggleFleetSelection(FleetRow row) {
-    toggleSelection(_selectionFleetId(row));
-  }
-
-  Iterable<String> _fleetSelectionIds(List<FleetRow> flat) =>
-      flat.map(_selectionFleetId);
-
-  /// Select-all header: from none or partial → select every row; from all → clear.
-  /// Does not rely on [Checkbox] tristate `next` (indeterminate taps may pass false).
-  void _onHeaderSelectAllTapped(List<FleetRow> flat) {
-    selectAllOrClear(_fleetSelectionIds(flat));
-  }
-
-  void _performCombine(List<FleetRow> flat) {
-    if (!_canCombineSelection(flat)) return;
-
-    final selected = Set<String>.from(selection.selectedIds);
-    final homeTransferRows = _homeTransferRows(flat, selected);
-    if (homeTransferRows != null &&
-        _isEligibleHomeTransferSource(homeTransferRows.source)) {
-      _openTransferToHomeDialog(
-        homeRow: homeTransferRows.home,
-        sourceRow: homeTransferRows.source,
-      );
-      return;
-    }
-    final targetId = _combineTargetFleetId(flat, selected);
-
-    FleetRow? targetRow;
-    for (final row in flat) {
-      if (_selectionFleetId(row) == targetId) {
-        targetRow = row;
-        break;
-      }
-    }
-    if (targetRow == null) return;
-
-    final targetFleet = _fleetForRow(targetRow);
-    if (targetFleet == null) return;
-
-    final mergedShips = <ShipInstance>[...targetFleet.ships];
-    for (final row in flat) {
-      final id = _selectionFleetId(row);
-      if (!selected.contains(id) || id == targetId) continue;
-      final f = _fleetForRow(row);
-      if (f != null) mergedShips.addAll(f.ships);
-    }
-
-    final merged = Fleet(
-      id: targetId,
-      ownerId: widget.humanPlayerId,
-      seaZoneId: targetFleet.seaZoneId,
-      inPortAtProvinceId: targetFleet.inPortAtProvinceId,
-      regionId: targetFleet.regionId,
-      ships: mergedShips,
-      mission: FleetMission.none,
-    );
-
-    final homeId = homeFleetIdFor(widget.humanPlayerId);
-    var updated = widget.game.worldState.fleets
-        .where((f) => !selected.contains(f.id))
-        .toList();
-    updated = [...updated, merged];
-    updated = updated
-        .where((f) => f.ships.isNotEmpty || f.id == homeId)
-        .toList();
-
-    final newGame = widget.game.copyWith(
-      worldState: widget.game.worldState.copyWith(fleets: updated),
-    );
-
-    clearSelection();
-    widget.bus.emit(NavalFleetsUpdatedEvent(game: newGame));
-  }
-
-  ({FleetRow home, FleetRow source})? _homeTransferRows(
-    List<FleetRow> flat,
-    Set<String> selectedIds,
-  ) {
-    if (selectedIds.length != 2) return null;
-    FleetRow? home;
-    FleetRow? source;
-    for (final row in flat) {
-      final id = _selectionFleetId(row);
-      if (!selectedIds.contains(id)) continue;
-      if (row.isHomeFleet) {
-        home = row;
-      } else {
-        source = row;
-      }
-    }
-    if (home == null || source == null) return null;
-    return (home: home, source: source);
-  }
-
-  String? _humanCapitalProvinceId() {
-    for (final p in widget.game.players) {
-      if (p.id == widget.humanPlayerId) return p.capitalProvinceId;
-    }
-    return null;
-  }
-
-  bool _provinceMatchesCapital(String provinceId, String capitalProvinceId) {
-    if (provinceId == capitalProvinceId) return true;
-    final capRegionId = ProvinceId.regionIdFrom(capitalProvinceId);
-    final capLocalId = ProvinceId.localIdFrom(capitalProvinceId);
-    return provinceId == capLocalId || provinceId == '$capRegionId|$capLocalId';
-  }
-
-  bool _seaZoneAdjacentToCapital({
-    required String sourceSeaZoneId,
-    required String sourceRegionId,
-    required String capitalProvinceId,
-  }) {
-    final capRegionId = ProvinceId.regionIdFrom(capitalProvinceId);
-    final capLocalId = ProvinceId.localIdFrom(capitalProvinceId);
-    final sourceSeaLocal = prefixedIdLocalSegment(sourceSeaZoneId);
-    final sourceSeaPrefixed = prefixedIdHasDelimiter(sourceSeaZoneId)
-        ? sourceSeaZoneId
-        : '$sourceRegionId|$sourceSeaZoneId';
-    final sourceSeaCandidates = <String>{
-      sourceSeaZoneId,
-      sourceSeaLocal,
-      sourceSeaPrefixed,
-    };
-    final capitalCandidates = <String>{
-      capitalProvinceId,
-      capLocalId,
-      '$capRegionId|$capLocalId',
-    };
-    for (final edge in widget.topology.edges) {
-      final a = edge.id1;
-      final b = edge.id2;
-      final aIsSea = sourceSeaCandidates.contains(a);
-      final bIsSea = sourceSeaCandidates.contains(b);
-      final aIsCap = capitalCandidates.contains(a);
-      final bIsCap = capitalCandidates.contains(b);
-      if ((aIsSea && bIsCap) || (bIsSea && aIsCap)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  bool _isEligibleHomeTransferSource(FleetRow sourceRow) {
-    final sourceFleet = _fleetForRow(sourceRow);
-    final capitalProvinceId = _humanCapitalProvinceId();
-    if (sourceFleet == null || capitalProvinceId == null) return false;
-    if (sourceFleet.ownerId != widget.humanPlayerId) return false;
-    if (!sourceFleet.isAtSea) {
-      final inPortId = sourceFleet.inPortAtProvinceId;
-      if (inPortId == null) return false;
-      return _provinceMatchesCapital(inPortId, capitalProvinceId);
-    }
-    final seaZoneId = sourceFleet.seaZoneId;
-    if (seaZoneId == null || seaZoneId.isEmpty) return false;
-    return _seaZoneAdjacentToCapital(
-      sourceSeaZoneId: seaZoneId,
-      sourceRegionId: sourceFleet.regionId,
-      capitalProvinceId: capitalProvinceId,
-    );
-  }
-
-  void _openTransferToHomeDialog({
-    required FleetRow homeRow,
-    required FleetRow sourceRow,
-  }) {
-    final homeFleet = _fleetForRow(homeRow);
-    final sourceFleet = _fleetForRow(sourceRow);
-    if (homeFleet == null || sourceFleet == null) return;
-    showDialog<void>(
-      context: context,
-      builder: (_) => TransferToHomeFleetDialog(
-        sourceFleet: sourceFleet,
-        homeFleet: homeFleet,
-        game: widget.game,
-        humanPlayerId: widget.humanPlayerId,
-        bus: widget.bus,
-      ),
-    );
-  }
-
   @override
   void didUpdateWidget(covariant NavalUnitsPanel oldWidget) {
     super.didUpdateWidget(oldWidget);
@@ -390,45 +153,6 @@ class _NavalUnitsPanelState extends BaseUnitsPanelState<NavalUnitsPanel> {
         _pendingScopedAutoCloseAfterMove = false;
       }
     }
-  }
-
-  void _openTrainDialog() {
-    widget.bus.closePanelThenEmit(OpenDialogEvent(trainNavalDialogId));
-  }
-
-  void _openSplitDialog(FleetRow row) {
-    final id = _selectionFleetId(row);
-    final fleet = widget.game.fleetById(id);
-    if (fleet == null) return;
-
-    final original = fleet;
-    showDialog<void>(
-      context: context,
-      builder: (ctx) => SplitFleetDialog(
-        originalFleet: original,
-        game: widget.game,
-        humanPlayerId: widget.humanPlayerId,
-        isHomeFleet: row.isHomeFleet,
-        bus: widget.bus,
-      ),
-    );
-  }
-
-  Future<void> _openMoveFleetDialog(FleetRow row) async {
-    if (row.isHomeFleet) return;
-    final fleet = widget.game.fleetById(row.fleetId);
-    final nonNullFleet = fleet;
-    if (nonNullFleet == null) return;
-    await showDialog<bool>(
-      context: context,
-      builder: (ctx) => MoveFleetDialog(
-        game: widget.game,
-        topology: widget.topology,
-        humanPlayerId: widget.humanPlayerId,
-        fleet: nonNullFleet,
-        bus: widget.bus,
-      ),
-    );
   }
 
   @override
